@@ -1,23 +1,61 @@
 package com.example.c196aloufi.Model;
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
+import androidx.annotation.NonNull;
+import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
+import java.time.LocalDate;
 import java.util.Date;
 
 @Entity(tableName = "courses")
 
-public class Courses {
+public class Courses implements Parcelable {
     @PrimaryKey (autoGenerate = true)
-    private int courseId;
+    private Integer courseId;
+    @ColumnInfo
     private String courseName;
+    @ColumnInfo
     private String instructorName;
+    @ColumnInfo
     private String instructorEmail;
+    @ColumnInfo
     private String instructorPhone;
+    @ColumnInfo
     private String courseStatus;
-    private Date startDate;
-    private Date endDate;
+    @ColumnInfo
+    private LocalDate startDate;
+    @ColumnInfo
+    private LocalDate endDate;
+    @ColumnInfo
     private String courseNote;
+
+    protected Courses(Parcel in) {
+        if (in.readByte() == 0) {
+            courseId = null;
+        }
+        courseName = in.readString();
+        instructorName = in.readString();
+        instructorEmail = in.readString();
+        instructorPhone = in.readString();
+        courseStatus = in.readString();
+        courseNote = in.readString();
+    }
+
+    public static final Creator<Courses> CREATOR = new Creator<Courses>() {
+        @Override
+        public Courses createFromParcel(Parcel in) {
+            return new Courses(in);
+        }
+
+        @Override
+        public Courses[] newArray(int size) {
+            return new Courses[size];
+        }
+    };
 
     @Override
     public String toString() {
@@ -34,11 +72,11 @@ public class Courses {
                 '}';
     }
 
-    public int getCourseId() {
+    public Integer getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(int courseId) {
+    public void setCourseId(Integer courseId) {
         this.courseId = courseId;
     }
 
@@ -82,19 +120,19 @@ public class Courses {
         this.courseStatus = courseStatus;
     }
 
-    public Date getStartDate() {
+    public LocalDate getStartDate() {
         return startDate;
     }
 
-    public void setStartDate(Date startDate) {
+    public void setStartDate(LocalDate startDate) {
         this.startDate = startDate;
     }
 
-    public Date getEndDate() {
+    public LocalDate getEndDate() {
         return endDate;
     }
 
-    public void setEndDate(Date endDate) {
+    public void setEndDate(LocalDate endDate) {
         this.endDate = endDate;
     }
 
@@ -106,8 +144,8 @@ public class Courses {
         this.courseNote = courseNote;
     }
 
-    public Courses(int courseId, String courseName, String instructorName, String instructorEmail, String instructorPhone,
-                   String courseStatus, Date startDate, Date endDate, String courseNote) {
+    public Courses(Integer courseId, String courseName, String instructorName, String instructorEmail, String instructorPhone,
+                   String courseStatus, LocalDate startDate, LocalDate endDate, String courseNote) {
         this.courseId = courseId;
         this.courseName = courseName;
         this.instructorName = instructorName;
@@ -117,5 +155,26 @@ public class Courses {
         this.startDate = startDate;
         this.endDate = endDate;
         this.courseNote = courseNote;
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(@NonNull Parcel dest, int flags) {
+        if (courseId == null) {
+            dest.writeByte((byte) 0);
+        } else {
+            dest.writeByte((byte) 1);
+            dest.writeInt(courseId);
+        }
+        dest.writeString(courseName);
+        dest.writeString(instructorName);
+        dest.writeString(instructorEmail);
+        dest.writeString(instructorPhone);
+        dest.writeString(courseStatus);
+        dest.writeString(courseNote);
     }
 }
