@@ -1,36 +1,65 @@
 package com.example.c196aloufi.Model;
 
-import java.util.Date;
+import android.os.Parcel;
+import android.os.Parcelable;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+import androidx.annotation.NonNull;
+import androidx.room.ColumnInfo;
 import androidx.room.Entity;
-import androidx.room.Ignore;
 import androidx.room.PrimaryKey;
 
 @Entity(tableName = "terms")
-public class Terms {
+public class Terms implements Parcelable {
 
     @PrimaryKey(autoGenerate = true)
-    private int termId;
-    private String termName;
-    private Date startDate;
-    private Date endDate;
+    public Integer termId;
+    @ColumnInfo(name = "termName")
+    public String termName;
+    @ColumnInfo(name = "startDate")
+    public LocalDate startDate;
+    @ColumnInfo(name = "endDate")
+    public LocalDate endDate;
 
+    protected Terms(Parcel in) {
+        if (in.readByte() == 0) {
+            termId = null;
+        } else {
+            termId = in.readInt();
+        }
+        termName = in.readString();
+    }
 
+    public static final Creator<Terms> CREATOR = new Creator<Terms>() {
+        @Override
+        public Terms createFromParcel(Parcel in) {
+            return new Terms(in);
+        }
+
+        @Override
+        public Terms[] newArray(int size) {
+            return new Terms[size];
+        }
+    };
+
+    @NonNull
     @Override
     public String toString() {
         return "Terms{" +
                 "termId=" + termId +
                 ", termName='" + termName + '\'' +
-                ", startDate=" + startDate +
-                ", endDate=" + endDate +
+                ", startDate=" + startDate.format(DateTimeFormatter.ISO_LOCAL_DATE) +
+                ", endDate=" + endDate.format(DateTimeFormatter.ISO_LOCAL_DATE) +
                 '}';
     }
 
-    public int getTermId() {
+    public Integer getTermId() {
         return termId;
     }
 
-    public void setTermId(int termId) {
+    public void setTermId(Integer termId) {
         this.termId = termId;
     }
 
@@ -42,26 +71,45 @@ public class Terms {
         this.termName = termName;
     }
 
-    public Date getStartDate() {
+    public LocalDate getStartDate() {
         return startDate;
     }
 
-    public void setStartDate(Date startDate) {
+    public void setStartDate(LocalDate startDate) {
         this.startDate = startDate;
     }
 
-    public Date getEndDate() {
+    public LocalDate getEndDate() {
         return endDate;
     }
 
-    public void setEndDate(Date endDate) {
+    public void setEndDate(LocalDate endDate) {
         this.endDate = endDate;
     }
 
-    public Terms(int termId, String termName, Date startDate, Date endDate) {
+    public Terms(Integer termId, String termName, LocalDate startDate, LocalDate endDate) {
         this.termId = termId;
         this.termName = termName;
         this.startDate = startDate;
         this.endDate = endDate;
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(@NonNull Parcel dest, int flags) {
+        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ISO_LOCAL_DATE;
+        if (termId == null) {
+            dest.writeByte((byte) 0);
+        } else {
+            dest.writeByte((byte) 1);
+            dest.writeInt(termId);
+        }
+        dest.writeString(termName);
+        dest.writeString(startDate.format(dateTimeFormatter));
+        dest.writeString(endDate.format(dateTimeFormatter));
     }
 }
