@@ -4,26 +4,51 @@ package com.example.c196aloufi.UserInterface;
 import android.app.AlertDialog;
 import android.app.DatePickerDialog;
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
 
 
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.c196aloufi.Database.AppRepo;
+import com.example.c196aloufi.Model.Terms;
 import com.example.c196aloufi.R;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.Calendar;
+import java.util.Date;
 
 public class AddTerm extends AppCompatActivity {
 
-    DatePickerDialog datePickerDialog;
+    Terms terms;
 
-    DatePickerDialog endDatePickerDialog;
+    String termTitle;
+
+    LocalDate startDate;
+
+    LocalDate endDate;
+
+    AppRepo appRepo;
+
+    EditText termTitleTxt;
 
     Button startDatePickerButton;
 
-    Button endDatePicker;
+    Button endDatePickerButton;
+
+    Button createBtn;
+
+    DatePickerDialog datePickerDialog;
+
+    Calendar myCal;
+
+    DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("MM-dd-yyyy") ;
 
 
     @Override
@@ -31,110 +56,79 @@ public class AddTerm extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_add_term);
 
-        initDatePicker();
+        appRepo = new AppRepo(getApplication());
+
+        setUpView();
+
+    }
+
+    private void setUpView() {
+        termTitleTxt = findViewById(R.id.termTitleTxt);
+        startDatePicker();
+        endDatePicker();
+        addNewTerm();
+    }
+
+
+    private void addNewTerm() {
+        createBtn = findViewById(R.id.createBtn);
+        createBtn.setOnClickListener(v -> {
+            termTitle = termTitleTxt.getText().toString();
+            if (isNull()) {
+                return;
+            }
+            Integer newId = appRepo.getAllTerms().get(appRepo.getAllTerms().size() - 1).getTermId() +1;
+            terms = new Terms(newId, termTitle, startDate, endDate);
+            appRepo.insert(terms);
+            Toast.makeText(AddTerm.this, "New Term Created.", Toast.LENGTH_SHORT).show();
+            Intent intent = new Intent(AddTerm.this, DetailedTerm.class);
+            startActivity(intent);
+        });
+    }
+
+    public boolean isNull() {
+        if (termTitleTxt.getText().toString().isEmpty()) {
+            Toast.makeText(this, "The new term must have a name.",Toast.LENGTH_SHORT).show();
+            return true;
+        }
+        else if (startDate == null) {
+            Toast.makeText(this, "The new term must have a start date.",Toast.LENGTH_SHORT).show();
+            return true;
+        }
+        else if (endDate == null) {
+            Toast.makeText(this, "The new term must have an end date.", Toast.LENGTH_SHORT).show();
+            return true;
+        }
+        else return false;
+    }
+    private void endDatePicker() {
+        endDatePickerButton = findViewById(R.id.endDatePickerButton);
+        endDatePickerButton.setOnClickListener(v -> {
+            myCal = Calendar.getInstance();
+            int day = myCal.get(Calendar.DAY_OF_MONTH);
+            int month = myCal.get(Calendar.MONTH);
+            int year = myCal.get(Calendar.YEAR);
+
+            datePickerDialog = new DatePickerDialog(AddTerm.this, (view, year1, month1, dayOfMonth) -> {
+                endDate = LocalDate.of(year1, (month1 +1), dayOfMonth);
+                endDatePickerButton.setText(dateTimeFormatter.format(endDate));
+            }, year, month, day);
+            datePickerDialog.show();
+        });
+    }
+    private void startDatePicker() {
         startDatePickerButton = findViewById(R.id.startDatePickerButton);
-        startDatePickerButton.setText(getTodaysDate());
-        initEndDatePicker();
-        endDatePicker = findViewById(R.id.endDatePicker);
-        endDatePicker.setText(getEndDate());
-    }
+        startDatePickerButton.setOnClickListener(v -> {
+            myCal = Calendar.getInstance();
+            int day = myCal.get(Calendar.DAY_OF_MONTH);
+            int month = myCal.get(Calendar.MONTH);
+            int year = myCal.get(Calendar.YEAR);
 
-    private String getTodaysDate() {
-        Calendar calendar = Calendar.getInstance();
-        int year = calendar.get(Calendar.YEAR);
-        int month = calendar.get(Calendar.MONTH);
-        int day = calendar.get(Calendar.DAY_OF_MONTH);
-        makeDateString(day, month, year);
-        return makeDateString(day, month, year);
-
-    }
-
-    private String getEndDate()  {
-        Calendar cal = Calendar.getInstance();
-        int year = cal.get(Calendar.YEAR);
-        int month = cal.get(Calendar.MONTH);
-        int day = cal.get(Calendar.DAY_OF_MONTH);
-        endDateString(day, month, year);
-        return endDateString(day, month, year);
-    }
-
-    private void initDatePicker() {
-        DatePickerDialog.OnDateSetListener dateSetListener = (view, year, month, dayOfMonth) -> {
-            String startDate = makeDateString(dayOfMonth, month, year);
-            startDatePickerButton.setText(startDate);
-        };
-        Calendar calendar = Calendar.getInstance();
-        int year = calendar.get(Calendar.YEAR);
-        int month = calendar.get(Calendar.MONTH);
-        int day = calendar.get(Calendar.DAY_OF_MONTH);
-        int style = AlertDialog.THEME_HOLO_DARK;
-        datePickerDialog = new DatePickerDialog(this, style, dateSetListener, year, month, day);
-    }
-
-    private void initEndDatePicker() {
-        DatePickerDialog.OnDateSetListener dateSetListener = (view, year, month, dayOfMonth) -> {
-            String endDate = endDateString(dayOfMonth, month, year);
-            endDatePicker.setText(endDate);
-        };
-        Calendar calendar = Calendar.getInstance();
-        int year = calendar.get(Calendar.YEAR);
-        int month = calendar.get(Calendar.MONTH);
-        int day = calendar.get(Calendar.DAY_OF_MONTH);
-        int style = AlertDialog.THEME_HOLO_DARK;
-        endDatePickerDialog = new DatePickerDialog(this, style, dateSetListener, year, month, day);
-    }
-    private String makeDateString(int dayOfMonth, int month, int year) {
-        month = month + 1;
-        return getDateFormat(month) + " " + dayOfMonth + " " + year;
-    }
-
-    private String endDateString (int dayOfMonth, int month, int year) {
-        month = month + 1;
-        return getDateFormat(month) + " " + dayOfMonth + " " + year;
-    }
-    private String getDateFormat(int month) {
-        if (month == 1)
-            return "Jan";
-        if (month == 2)
-            return "Feb";
-        if (month == 3)
-            return "Mar";
-        if (month == 4)
-           return "Apr";
-        if (month == 5)
-            return "May";
-        if (month == 6)
-            return "Jun";
-        if (month == 7)
-            return "Jul";
-        if (month == 8)
-            return "Aug";
-        if (month == 9)
-            return "Sep";
-        if (month == 10)
-            return "Oct";
-        if (month == 11)
-            return "Nov";
-        if (month ==12)
-            return "Dec";
-        return "Jan";
-    }
-
-    public void addTermBackBtn(View view) {
-        Intent intent = new Intent(AddTerm.this, DetailedTerm.class);
-        startActivity(intent);
-    }
-
-    public void addTermCreateBtn(View view) {
-        Intent intent = new Intent(AddTerm.this, DetailedTerm.class);
-        startActivity(intent);
-    }
-
-    public void openStartDate(View view) {
-        datePickerDialog.show();
-    }
-
-    public void openEndDate(View view) {
-        endDatePickerDialog.show();
+            datePickerDialog = new DatePickerDialog(AddTerm.this, (view, year1, month1, dayOfMonth) -> {
+                startDate = LocalDate.of(year1, (month1 +1), dayOfMonth);
+                startDatePickerButton.setText(dateTimeFormatter.format(startDate));
+            }, year, month, day);
+            datePickerDialog.show();
+        });
     }
 }
